refactored to metrics pkg
