+++ cmd/clusters-service/app/options.go
@@ -3,7 +3,7 @@ package app
 import (
 	"github.com/go-logr/logr"
 	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
-	"github.com/weaveworks/weave-gitops-enterprise/cmd/clusters-service/pkg/server"
+	"github.com/weaveworks/weave-gitops-enterprise/pkg/metrics"
 	"github.com/weaveworks/weave-gitops-enterprise/pkg/query/collector"
 
 	"github.com/weaveworks/weave-gitops-enterprise/cmd/clusters-service/pkg/git"
@@ -55,7 +55,7 @@ type Options struct {
 	UIConfig                  string
 	PipelineControllerAddress string
 	CollectorServiceAccount   collector.ImpersonateServiceAccount
-	MetricsServerConf         server.MetricsServerConf
+	MetricsOptions            metrics.Options
 }
 
 type Option func(*Options)
@@ -271,10 +271,10 @@ func WithCollectorServiceAccount(name, namespace string) Option {
 // WithMetrics configures prometheus metrics
 func WithMetrics(enabled bool, address string, log logr.Logger) Option {
 	return func(o *Options) {
-		o.MetricsServerConf = server.MetricsServerConf{
-			Enabled: enabled,
-			Address: address,
-			Log:     log,
+		o.MetricsOptions = metrics.Options{
+			Enabled:       enabled,
+			ServerAddress: address,
+			Log:           log,
 		}
 	}
 }
