+++ pkg/metrics/metrics.go
@@ -1,0 +1,9 @@
+package metrics
+
+import "github.com/go-logr/logr"
+
+type Options struct {
+	Enabled       bool
+	ServerAddress string
+	Log           logr.Logger
+}
