+++ pkg/metrics/prometheus.go
@@ -1,7 +1,6 @@
-package server
+package metrics
 
 import (
-	"github.com/go-logr/logr"
 	prom2 "github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/slok/go-http-metrics/metrics/prometheus"
@@ -10,20 +9,14 @@ import (
 	"net/http"
 )
 
-type MetricsServerConf struct {
-	Enabled bool
-	Address string
-	Log     logr.Logger
-}
-
-// NewMetricsServer creates and starts a prometheus metrics server in /metrics path
+// NewPrometheusServer creates and starts a prometheus metrics server in /metrics path
 // with the gatherers and configuration given as argument.
-func NewMetricsServer(conf MetricsServerConf, gatherers prom2.Gatherers) *http.Server {
-	log := conf.Log.WithName("metrics-server")
+func NewPrometheusServer(opts Options, gatherers prom2.Gatherers) *http.Server {
+	log := opts.Log.WithName("metrics-server")
 	metricsMux := http.NewServeMux()
 	metricsMux.Handle("/metrics", promhttp.HandlerFor(gatherers, promhttp.HandlerOpts{}))
 	metricsServer := &http.Server{
-		Addr:    conf.Address,
+		Addr:    opts.ServerAddress,
 		Handler: metricsMux,
 	}
 
@@ -37,9 +30,9 @@ func NewMetricsServer(conf MetricsServerConf, gatherers prom2.Gatherers) *http.S
 	return metricsServer
 }
 
-// WithMetrics instruments http server with a prometheus metrics filter to generate
+// WithHttpMetrics instruments http server with a prometheus metrics filter to generate
 // golden signal metrics using https://github.com/slok/go-http-metrics
-func WithMetrics(h http.Handler) http.Handler {
+func WithHttpMetrics(h http.Handler) http.Handler {
 	recorder := prometheus.NewRecorder(prometheus.Config{})
 
 	metricsMiddleware := middleware.New(middleware.Config{
